Fix warnings when building Pluto as 32-bit

+++ src/lparser.cpp
@@ -1988,7 +1988,7 @@ static void primaryexp (LexState *ls, expdesc *v) {
       if (strcmp(ls->t.seminfo.ts->contents, "values") == 0) {
         luaX_next(ls);
         checknext(ls, '('); checknext(ls, ')');
-        const EnumDesc* ed = &ls->enums.at(v->u.ival);
+        const EnumDesc* ed = &ls->enums.at((size_t)v->u.ival);
         size_t i = 0;
         newtable(ls, v, [ed, &i](expdesc *e) {
           if (i == ed->enumerators.size())
@@ -2001,7 +2001,7 @@ static void primaryexp (LexState *ls, expdesc *v) {
       else if (strcmp(ls->t.seminfo.ts->contents, "names") == 0) {
         luaX_next(ls);
         checknext(ls, '('); checknext(ls, ')');
-        const EnumDesc* ed = &ls->enums.at(v->u.ival);
+        const EnumDesc* ed = &ls->enums.at((size_t)v->u.ival);
         size_t i = 0;
         newtable(ls, v, [ed, &i](expdesc *e) {
           if (i == ed->enumerators.size())
@@ -2014,7 +2014,7 @@ static void primaryexp (LexState *ls, expdesc *v) {
       else if (strcmp(ls->t.seminfo.ts->contents, "kvmap") == 0) {
         luaX_next(ls);
         checknext(ls, '('); checknext(ls, ')');
-        const EnumDesc* ed = &ls->enums.at(v->u.ival);
+        const EnumDesc* ed = &ls->enums.at((size_t)v->u.ival);
         size_t i = 0;
         newtable(ls, v, [ed, &i](expdesc *key, expdesc *val) {
           if (i == ed->enumerators.size())
@@ -2030,7 +2030,7 @@ static void primaryexp (LexState *ls, expdesc *v) {
       else if (strcmp(ls->t.seminfo.ts->contents, "vkmap") == 0) {
         luaX_next(ls);
         checknext(ls, '('); checknext(ls, ')');
-        const EnumDesc* ed = &ls->enums.at(v->u.ival);
+        const EnumDesc* ed = &ls->enums.at((size_t)v->u.ival);
         size_t i = 0;
         newtable(ls, v, [ed, &i](expdesc *key, expdesc *val) {
           if (i == ed->enumerators.size())
